+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 2020-09-03
+## 2020-04-09
 - Extended the question to identify cases related to the coronavirus outbreak to:
   - Allegations
   - Enquiries
   - Projects
+
+## 2020-04-07
+- Added ability to edit the coronavirus related flag
 
 ## 2020-04-03
 - Added a question for MSA users to identify cases related to the coronavirus outbreak.